% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{cb}
\alias{cb}
\title{Codebook}
\usage{
cb(df)
}
\arguments{
\item{df}{`df` should be a data frame.}
}
\value{
`cb(df)` returns a data frame.
}
\description{
`cb` provides summary statistics of a data frame,
just like the `codebook, compact` command in Stata.
}
\examples{
library(dplyr)
abdata = haven::read_dta("http://www.stata-press.com/data/r13/abdata.dta")
cb(nlsw88)

}

