% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{clusterCov}
\alias{clusterCov}
\title{Generate cluster covariance plots}
\usage{
clusterCov(x)
}
\arguments{
\item{x}{an object of type hmc of clusterCov}
}
\value{
a cluster covariance plot
}
\description{
Generate cluster covariance plots
}
\examples{
dat <- iris[, -5]
truth <- iris[, 5]
L <- hmc(dat, truth = truth, modelNames = c("VVV"))
clusterCov(L$dat$sigma)
}
