% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colStats.R
\name{colStats}
\alias{colStats}
\title{Show summary of data and feature types}
\usage{
colStats(dat)
}
\arguments{
\item{dat}{data}
}
\description{
Show summary of data and feature types
}
\examples{
dat0 <- iris[, -5]
dat1 <- data.frame(matrix(as.integer(rbinom(150*5, 10, 0.5)), nrow=150,ncol=5))
z <- complex(real = stats::rnorm(150), imaginary = stats::rnorm(150))
dat <- data.frame(dat0, dat1, let = paste0(letters, 1:150), cplx = z,stringsAsFactors = FALSE)
cs <- colStats(dat0)
plot(cs)
}
