% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{comp}
\alias{comp}
\title{Data compression}
\usage{
comp(dat, dir = 2, nmethod = "samp", dmethod = "samp", nnum = 1000,
  dnum = 100)
}
\arguments{
\item{dat}{data}

\item{dir}{direction in which to compress, '1' indicates rows, '2'
indicates columns, and '3' performs columns then rows.}

\item{nmethod}{a string specifying the method to use for compressing
rows.}

\item{dmethod}{a string specifying the method to use for compressing
columns}

\item{nnum}{a number indicating desired sample size.}

\item{dnum}{a number indicating desired dimension.}
}
\value{
a compressed version of the input data.
}
\description{
Data compression
}

