% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{Describe a data frame}
\usage{
d(df, compact = TRUE)
}
\arguments{
\item{df}{`df` should be a data frame.
`d` works best with labelled data frames (from the `haven` package).}

\item{compact}{`compact` should be logical.}
}
\value{
`d(df)` returns a data frame that contains metadata records
for the dataset.
}
\description{
`d` describes a data frame, just like the `d` command in Stata.
}
\examples{
nlsw88 = haven::read_dta("http://www.stata-press.com/data/r13/nlsw88.dta")
d(nlsw88)

}

