% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{d1heat}
\alias{d1heat}
\title{Generate 1d heatmaps}
\usage{
d1heat(dat, breaks = "Scott", ccol = "black", trunc = c(0.01, 0.99))
}
\arguments{
\item{dat}{the data}

\item{breaks}{see \code{\link[graphics]{hist}}}

\item{ccol}{colors for features.}

\item{trunc}{numeric vector of length 2 to truncate bottom and top
respectively, defaults to c(0.01,0.99), set to NA for no truncation.}
}
\value{
an object of type d1heat
}
\description{
Generate 1d heatmaps
}
\details{
For each feature column a 1D heatmap is generated and
plotted as a geom_tile object.
}
\examples{
dat <- iris[, -5]
d <- d1heat(dat, breaks = "Scott", ccol = 1:4, trunc = c(0.01,0.99))
plot(d)
plot(d, mycol = colorpanel(255, "white", "blue"))
}
