% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{genHTML}
\alias{genHTML}
\title{Generate an Rmarkdown with exploratory plots}
\usage{
genHTML(x, outfile, use.plotly = TRUE, scale = TRUE)
}
\arguments{
\item{x}{data}

\item{outfile}{A string denoting the location to put the output html
file.}

\item{use.plotly}{A Boolean to specify if ggplotly is used.}

\item{scale}{A Boolean to specify whether column-wise scaling is
performed before analysis.}
}
\value{
An html document.
}
\description{
Generate an Rmarkdown with exploratory plots
}
\details{
Generates an html file of various exploratory plots via
RMarkdown. "There is no excuse for failing to plot and look." ~ J. W.
Tukey (Exploratory Data Analysis, p 43)
}
\examples{
require(meda)
set.seed(73)
n <- 100
p <- 48
x <- data.frame(matrix(runif(n*p) + rnorm(n*p, sd = 0.01), nrow = n, ncol = p))
colnames(x)  <- state.name[1:dim(x)[2]]

outfile <- "./ex.html"
use.plotly <- FALSE; scale <- TRUE
x <- iris[, -5]
genHTML(x, outfile, use.plotly = FALSE, scale = TRUE)


}
\seealso{
\code{\link[rmarkdown]{render}}
}

