% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc}
\alias{hmc}
\title{Hierarchical MCLUST}
\usage{
hmc(dat, maxit = 100)
}
\arguments{
\item{dat}{A data frame or matrix
file.}

\item{maxit}{maximum number of levels in the hierarchy.}
}
\value{
A vector of cluster labels
}
\description{
Hierarchical MCLUST
}
\details{
Performs hierarchical mclust for k (G) = 1 or 2.
If bic yields k = 2, then clustering is performed again.
If bic yields k = 1, then that cluster is frozen.
}
\examples{
library(meda)
Z <- rep(1:5, each = 50)
mu <- c(-10,-5,5,10,20)
set.seed(391857)
test <- rnorm(length(Z), mean = mu[Z], sd = 1)
L <- hmc(test, maxit = Inf)
plot(cbind(test,0), col = Z)
for(i in 2:dim(L)[2]){
 points(cbind(test,i/dim(L)[2]), col=L[,i])
 }

mu2d <- matrix(3*c(1,0,-1,0,-2,0,1,0,-1,-2), 5,2)
set.seed(1234)
test2d <- data.frame(
            x = rnorm(length(Z), mean = mu2d[Z,1], sd = 1),
            y = rnorm(length(Z), mean = mu2d[Z,2], sd = 1))

L <- hmc(test2d, maxit = 100)
plot(test2d, col = L[, dim(L)[2]], pch = Z)

}

