% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcTree.R
\name{hmcTree}
\alias{hmcTree}
\title{Generate a binary hierarchical mclust tree}
\usage{
hmcTree(dat, maxDepth = 5, modelNames = c("VVV"))
}
\arguments{
\item{dat}{a data matrix}

\item{maxDepth}{maximum tree depth, defaults to 5}

\item{modelNames}{passed to mclust, defaults to
unrestricted GMM i.e. "VVV"}
}
\value{
binary hierarchical mclust classification output
}
\description{
Generate a binary hierarchical mclust tree
}
\details{
BIC is run for k = {1,2}, if k = 2 then each node is
propagated down the tree.  If k = 1, then that node is frozen.
}
\examples{
dat <- iris[, -5]
truth <- iris[, 5]
L <- hmcTree(dat)
}
