% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmpp.r
\name{kmpp}
\alias{kmpp}
\title{K-Means++ initialization to K-means}
\usage{
kmpp(x, k = 2)
}
\arguments{
\item{x}{The data formatted as a matrix.}

\item{k}{The number of clusters desired. \code{k < nrows(x) || length(x)}}
}
\value{
An object of class `"kmeans"'
}
\description{
K-Means++ initialization to K-means
}
\details{
A set of centers are calculated as suggested by
arthur and Vassilvitskii and used to initialize the kmeans algorithm.
The first center point is chosen at random.  Each subsequent center
is chosen with probability P = D(x')^2/( ∑_{x\in X} D(x)^2) (see refernces)
}
\examples{
set.seed(13)
z <- c(0,6,18,24)[sample(4, 1e2, replace=TRUE)]
x <- rnorm(1e2, mean = z, sd = 1)
kdf <- kmpp(x, 4)
plot(x, rep(1,length(x)), col = kdf$cluster)

z2 <- cbind(c(0,8,0,8),c(0,5,5,0))[sample(4,1e6, replace = TRUE),]
xy <- data.frame(x = rnorm(1e3, mean = z2[, 1]), y = rnorm(1e3, mean = z2[, 2]))
kdf <- kmpp(xy, 4)
kdf <- kmeans(xy, centers = 4)
plot(xy, col = kdf$cluster)

}
\references{
: \code{@inproceedings{arthur2007,
  title={k-means++: The advantages of careful seeding},
  author={Arthur, David and Vassilvitskii, Sergei},
  booktitle={Proceedings of the eighteenth annual ACM-SIAM symposium on Discrete algorithms},
  pages={1027--1035},
  year={2007},
  organization={Society for Industrial and Applied Mathematics}
}}
}
\seealso{
\code{\link{kmeans}}
}

