% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{l}
\alias{l}
\title{List observations}
\usage{
l(df, by = NULL, n = 20)
}
\arguments{
\item{df}{`df` should be a data frame.}

\item{by}{`by` should be a string indicating variables in the data frame.}

\item{n}{`n` should be a positive integer.}
}
\value{
`l(df)` returns a data frame.
}
\description{
`l` lists observations in a data frame, just like the `l` command in Stata.
}
\examples{
library(dplyr)
abdata = haven::read_dta("http://www.stata-press.com/data/r13/abdata.dta")
abdata \%>\% select(id, year, wage) \%>\% l("id", n = 30)

}

