% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookfor.R
\name{lookfor}
\alias{lookfor}
\title{Look For Variables}
\usage{
lookfor(df, kw)
}
\arguments{
\item{df}{The input `df` should be a data frame.
`lookfor` works best with labelled data frames (from the `haven` package).}

\item{kw}{The keyword to search, `kw`, should be a string of length 1.}
}
\value{
`lookfor(df, kw)` returns a data frame that contains rows which
match the keyword being searched.
}
\description{
`lookfor` searches for string in variable names and labels,
just like the `lookfor` command in Stata.
}
\examples{
nlsw88 = haven::read_dta("http://www.stata-press.com/data/r13/nlsw88.dta")
lookfor(nlsw88, "mar") # search for variables that are related to marriage

}

