% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondMoment.R
\name{medacor}
\alias{medacor}
\title{Generate a correlation plot}
\usage{
medacor(dat, ccol = "black")
}
\arguments{
\item{dat}{data}

\item{ccol}{colors for column labels}
}
\value{
A correlation plot
}
\description{
Generate a correlation plot
}
\examples{
dat <- iris[, -5]
ccol <- c("darkgreen", "red", "green", "red")
p1 <- medacor(dat, ccol)
}
