% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.1dheat}
\alias{p.1dheat}
\title{Generate 1d heatmaps}
\usage{
p.1dheat(dat, breaks = "Scott")
}
\arguments{
\item{dat}{the data}

\item{breaks}{see \code{\link[graphics]{hist}}}
}
\value{
a ggplot object
}
\description{
Generate 1d heatmaps
}
\details{
For each feature column a 1D heatmap is generated and
plotted as a geom_tile object.
}
\examples{
dat <- iris[, -5]
p.1dheat(dat)
p.1dheat(dat) + coord_flip()

}

