% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.clusterMeans}
\alias{p.clusterMeans}
\title{Generate cluster parameter plots}
\usage{
p.clusterMeans(mod)
}
\arguments{
\item{mod}{mclust output}
}
\value{
heatmap and line plot of cluster means
}
\description{
Generate cluster parameter plots
}
\examples{
dat <- iris[, -5]
out <- p.bic(dat)
truth <- iris[, 5]
tryCatch(md1 <- p.mclust(out$dat, out$bicO, truth = truth))
mod <- md1 
p <- p.clusterMeans(mod)
grid.arrange(p[[1]], p[[2]], ncol = 2)
}

