% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.cor}
\alias{p.cor}
\title{Generate a correlation plot}
\usage{
p.cor(dat, nmethod = "samp", dmethod = "samp")
}
\arguments{
\item{dat}{data}

\item{nmethod}{a string specifying the method to use for compressing
rows.}

\item{dmethod}{a string specifying the method to use for compressing
columns}
}
\value{
A correlation plot
}
\description{
Generate a correlation plot
}

