% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.cumvar}
\alias{p.cumvar}
\title{Generate a cumulative variance plot}
\usage{
p.cumvar(dat)
}
\arguments{
\item{dat}{data}
}
\value{
A correlation plot
}
\description{
Generate a cumulative variance plot
}
\details{
Uses getElbows from
\url{http://www.cis.jhu.edu/~parky/Synapse/getElbows.R}
}
\examples{
dat <- iris[, -5]
p.cumvar(dat)
}

