% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.dend}
\alias{p.dend}
\title{Generate dendrogram from clustering}
\usage{
p.dend(tree)
}
\arguments{
\item{tree}{hmcTree output}
}
\value{
dendrogram plot
}
\description{
Generate dendrogram from clustering
}

