% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.eig}
\alias{p.eig}
\title{Generate singular vector plots}
\usage{
p.eig(dat)
}
\arguments{
\item{dat}{data}
}
\value{
pairs plot and heatmap of right singular vectors
}
\description{
Generate singular vector plots
}
\examples{
dat <- iris[, -5]
p1 <- p.eig(dat)
## grid.arrange(p1[[1]], p1[[2]], nrow = 2)
p1[[1]]
p1[[2]]
}

