% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.energy}
\alias{p.energy}
\title{Generate an energy distance heatmap}
\usage{
p.energy(dat, colCol = NULL)
}
\arguments{
\item{dat}{data}

\item{colCol}{colors for column labels}
}
\value{
A correlation plot
}
\description{
Generate an energy distance heatmap
}
\details{
Takes a random sample of 1000 rows and computes
  energy statistics with p-values.
}

