% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.heat}
\alias{p.heat}
\title{Generate a heatmap plot}
\usage{
p.heat(dat, use.plotly)
}
\arguments{
\item{dat}{data}

\item{use.plotly}{Boolean to use plotly}
}
\value{
A plot
}
\description{
Generate a heatmap plot
}

