% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.hmc}
\alias{p.hmc}
\title{Generate binary hierarchical mclust tree}
\usage{
p.hmc(dat, truth = NULL, maxDim = Inf, maxDepth = 6, modelNames = NULL)
}
\arguments{
\item{dat}{data}

\item{truth}{true labels if any}

\item{maxDim}{maximum dimensions to plot}

\item{maxDepth}{maximum tree depth}

\item{modelNames}{model names for mclust see \code{\link[mclust]{mclustModelNames}}}
}
\value{
binary hierarchical mclust classification output
}
\description{
Generate binary hierarchical mclust tree
}
\details{
BIC is run for k = {1,2}, if k = 2 then each node is
propagated down the tree.  If k = 1, then that node is frozen. 
If a singleton exists, the level takes a step back.
}
\examples{
dat <- iris[, -5]
truth <- NULL #iris[,5]
L <- p.hmc(dat, truth = truth, modelNames = c("VVV"), maxDim = 6)
}

