% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.hmclust}
\alias{p.hmclust}
\title{Generate binary hierarchical mclust output}
\usage{
p.hmclust(dat, truth = NULL, maxd = Inf)
}
\arguments{
\item{dat}{data}

\item{truth}{true labels if any}

\item{maxd}{maximum dimensions to plot}
}
\value{
binary hierarchical mclust classification output
}
\description{
Generate binary hierarchical mclust output
}
\details{
BIC is run for k = {1,2}, if k = 2 then each node is
propagated down the tree.  If k = 1, then that node is frozen. 
If a singleton exists, the level takes a step back.
}
\examples{
dat <- iris[, -5]
truth <- iris[,5]
hmcL <- p.hmclust(dat, truth)
L <- p.hmclust(dat)
}

