% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.jitter}
\alias{p.jitter}
\title{Generate jittered scatter plots colored by class / cluster}
\usage{
p.jitter(dat, clusterLab = NULL)
}
\arguments{
\item{dat}{data}

\item{clusterLab}{true or predicted labels if any}
}
\value{
A jittered scatter plot
}
\description{
Generate jittered scatter plots colored by class / cluster
}
\examples{
dat <- iris[, -5]
clusterLab <- iris[, 5]
p.jitter(dat, clusterLab)
}

