% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.location}
\alias{p.location}
\title{Generate location estimates}
\usage{
p.location(dat, ccol = NULL)
}
\arguments{
\item{dat}{data}

\item{ccol}{colors for features.}
}
\value{
a list of ggplot objects, see details.
}
\description{
Generate location estimates
}
\examples{
dat <- iris[, -5]
l1 <- p.location(dat)
l2 <- p.location(dat, ccol)
grid.arrange(l1[[1]],l1[[2]], ncol = 2)
grid.arrange(l2[[1]],l2[[2]], ncol = 2)
}

