% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.mclust}
\alias{p.mclust}
\title{Generate mclust output}
\usage{
p.mclust(dat, bic)
}
\arguments{
\item{dat}{data that p.bic has been run on}

\item{bic}{output from p.bic or \code{\link[mclust]{mclustBIC}}}
}
\value{
mclust classification output
}
\description{
Generate mclust output
}
\examples{
out <- p.bic(iris[, -5])
p.mclust(out$dat, out$bicO)
}

