% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.outlier}
\alias{p.outlier}
\title{Generate an outlier plot}
\usage{
p.outlier(dat, alev = 0.01)
}
\arguments{
\item{dat}{data}

\item{alev}{alpha level}
}
\value{
An outlier plot
}
\description{
Generate an outlier plot
}

