% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.rsv}
\alias{p.rsv}
\title{Generate singular vector plots}
\usage{
p.rsv(dat, ccol = "black", maxd = Inf)
}
\arguments{
\item{dat}{data}

\item{ccol}{feature colors}

\item{maxd}{maximum dimensions to plot}
}
\value{
pairs plot and heatmap of right singular vectors
}
\description{
Generate singular vector plots
}
\examples{
dat <- iris[, -5]
p1 <- p.rsv(dat, ccol = c("black", "black", "blue", "blue"))
## grid.arrange(p1[[1]], p1[[2]], nrow = 2)
p1[[1]]
p1[[2]]
}

