% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.stackM}
\alias{p.stackM}
\title{Generate stacked level mean plot}
\usage{
p.stackM(tree)
}
\arguments{
\item{tree}{the hmcTree object}
}
\value{
a stacked level mean plot
}
\description{
Generate stacked level mean plot
}

