% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.try}
\alias{p.try}
\title{Try to plot data}
\usage{
p.try(FUN, dat, use.plotly = NULL)
}
\arguments{
\item{FUN}{a p.* function from meda::}

\item{dat}{data}
}
\value{
The output of FUN or an error message.
}
\description{
Try to plot data
}

