% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.violin}
\alias{p.violin}
\title{Generate violin/jitter plot of data}
\usage{
p.violin(dat, use.plotly, dmethod = "cur")
}
\arguments{
\item{dat}{data}

\item{use.plotly}{Boolean to use plotly}

\item{dmethod}{a string specifying the method to use for compressing
columns if d > 30.}
}
\value{
A jitter/violin plot
}
\description{
Generate violin/jitter plot of data
}
\seealso{
\code{\link{comp}}
}

