% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsHex.R
\name{pairhex}
\alias{pairhex}
\title{Generate a pairs hex binned plot}
\usage{
pairhex(dat, maxd = Inf, colramp = BTC, ccol = "black", loess = TRUE,
  lmline = TRUE)
}
\arguments{
\item{dat}{data}

\item{maxd}{maximum dimensions to plot}

\item{colramp}{Color Ramp used, BTC or BTY}

\item{ccol}{colors for features}

\item{loess}{boolean for loess curve}

\item{lmline}{boolean for lm line}
}
\value{
A lattice splom
}
\description{
Generate a pairs hex binned plot
}
\examples{
dat <- iris[,-5]
pairhex(dat, colramp = magent, ccol = c("blue", "red"))
}
