% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{plot.clusterCov}
\alias{plot.clusterCov}
\title{Generate cluster covariance/correlation plots}
\usage{
\method{plot}{clusterCov}(x, ...)
}
\arguments{
\item{x}{object of type hmc}

\item{...}{used for 'ccol' feature colors}
}
\value{
heatmaps of cluster covariances.
}
\description{
Generate cluster covariance/correlation plots
}
\examples{
dat <- iris[, -5]
truth <- iris[, 5]
L <- hmc(dat, truth = truth, modelNames = c("VVV"))
plot(L, ccol = 1:4)
clusterCov(L)
}
