% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumvar.R
\name{plot.cumvar}
\alias{plot.cumvar}
\title{Plot method for cumulative variance plots}
\usage{
\method{plot}{cumvar}(x, ...)
}
\arguments{
\item{x}{an object of type cumvar}

\item{...}{unused}
}
\value{
A cumulative variance plot
}
\description{
Plot method for cumulative variance plots
}
