% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcTree.R
\name{plot.hmc}
\alias{plot.hmc}
\title{Generate binary hierarchical mclust tree plot}
\usage{
\method{plot}{hmc}(x, ...)
}
\arguments{
\item{x}{an object of type hmc}

\item{...}{plotDend Boolean for dendrogram plot and maxd for max
plotting dimension}
}
\description{
Generate binary hierarchical mclust tree plot
}
