% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondMoment.R
\name{plot.medacor}
\alias{plot.medacor}
\title{Generate correlation plots}
\usage{
\method{plot}{medacor}(x, ...)
}
\arguments{
\item{x}{an object of type medacor}

\item{...}{unused}
}
\value{
a corrplot
}
\description{
Generate correlation plots
}
