% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{plot.mlocation}
\alias{plot.mlocation}
\title{Generate location estimate plots}
\usage{
\method{plot}{mlocation}(x, ...)
}
\arguments{
\item{x}{an object of type location}

\item{...}{unused}
}
\value{
a list of ggplot objects.
}
\description{
Generate location estimate plots
}
