% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{plot.outliers}
\alias{plot.outliers}
\title{Generate an outlier plot}
\usage{
\method{plot}{outliers}(x, ...)
}
\arguments{
\item{x}{an object of type outliers}

\item{...}{unused}
}
\value{
An outlier plot
}
\description{
Generate an outlier plot
}
