% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralAnalysis.R
\name{plot.rsv}
\alias{plot.rsv}
\title{Generate singular vector data}
\usage{
\method{plot}{rsv}(x, ...)
}
\arguments{
\item{x}{an object of type rsv}

\item{...}{used for ccol feature colors}
}
\value{
pairs plot and heatmap of right singular vectors
}
\description{
Generate singular vector data
}
