% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcTree.R
\name{plotDend}
\alias{plotDend}
\title{plot a dendrogram from an hmc object}
\usage{
plotDend(dl)
}
\arguments{
\item{dl}{an hmc object}
}
\value{
a dendrogram plot
}
\description{
plot a dendrogram from an hmc object
}
