% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralAnalysis.R
\name{rsv}
\alias{rsv}
\title{Generate singular vector data}
\usage{
rsv(dat, ccol = "black", maxd = Inf)
}
\arguments{
\item{dat}{data}

\item{ccol}{feature colors}

\item{maxd}{maximum dimensions to plot}
}
\value{
pairs plot and heatmap of right singular vectors
}
\description{
Generate singular vector data
}
\examples{
dat <- iris[, -5]
p1 <- rsv(dat, ccol = 2:5)

}
