% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{stackM}
\alias{stackM}
\title{Generate stacked level mean plot}
\usage{
stackM(hmcL, ccol = "black", centered = FALSE, maxDepth = Inf,
  depth = 5)
}
\arguments{
\item{hmcL}{object of type hmc}

\item{ccol}{colors for feature labels}

\item{centered}{boolen that skips level one if data was centered}

\item{maxDepth}{maximum number of levels to plot}

\item{depth}{forces via copying levels to show down to depth.}
}
\value{
a stacked level mean plot
}
\description{
Generate stacked level mean plot
}
\examples{
dat <- as.data.frame(scale(iris[, -5], center = TRUE, scale = TRUE))
hmcL <- hmc(dat, modelNames = "VVV")
plot(hmcL, plotDend = TRUE)
stackM(hmcL, centered = TRUE)
centered = TRUE
maxDepth = Inf
depth = 5
}
