% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{Describe a data frame}
\usage{
d(df)
}
\arguments{
\item{df}{\code{df} should be a data frame.
\code{d} works best with labelled data frames (from the \code{haven} package).}
}
\value{
\code{d(df)} returns a data frame that contains metadata records
for the dataset.
}
\description{
\code{d} describes a data frame, just like the \code{d} command in Stata.
}
\examples{
nlsw88 = haven::read_dta("http://www.stata-press.com/data/r13/nlsw88.dta")
d(nlsw88)
}

