% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1heat.R
\name{d1heat}
\alias{d1heat}
\title{Generate 1d heatmaps}
\usage{
d1heat(dat, breaks = "Scott", ccol = "black")
}
\arguments{
\item{dat}{the data}

\item{breaks}{see \code{\link[graphics]{hist}}}

\item{ccol}{colors for features.}
}
\value{
an object of type d1heat
}
\description{
Generate 1d heatmaps
}
\details{
For each feature column a 1D heatmap is generated and
plotted as a geom_tile object.
}
\examples{
dat <- iris[, -5]
d1 <- d1heat(dat, breaks = 14, ccol = 1:4)
plot(d1)
}
