% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{genHTML}
\alias{genHTML}
\title{Generate an Rmarkdown with exploratory plots}
\usage{
genHTML(x, outfile, use.plotly = TRUE, scale = TRUE, samp = 10000)
}
\arguments{
\item{x}{data}

\item{outfile}{A string denoting the location to put the output html
file.}

\item{use.plotly}{A Boolean to specify if ggplotly is used.}

\item{scale}{A Boolean to specify whether column-wise scaling is
performed before analysis.}

\item{samp}{An integer specifying the size of the random sample to be
taken if the number of points exceeds 1e5.}
}
\value{
An html document via RMarkdown.
}
\description{
Generate an Rmarkdown with exploratory plots
}
\details{
Generates an html file of various exploratory plots via
RMarkdown. "There is no excuse for failing to plot and look." ~ J. W.
Tukey (Exploratory Data Analysis, p 43)
}
\examples{
require(meda)
x <- iris[, -5]
y <- data.frame(matrix(rnorm(1e3*24, mean = rep(c(1,4,2,5), each =
1e3), sd = 0.25), ncol = 24))
outfile.1 <- paste0(getwd(), '/ex1.html')
outfile.2 <- paste0(getwd(), '/ex2.html')
use.plotly <- FALSE
scale <- TRUE
print("Now run 'genHTML(x, outfile, use.plotly, scale)'")
\dontrun{
genHTML(x, outfile.1, use.plotl = TRUE, scale)
genHTML(W, outfile, use.plotl = TRUE, scale)
}

}
\seealso{
\code{\link[rmarkdown]{render}}
}

