% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{genHTML}
\alias{genHTML}
\title{Generate Rmarkdown}
\usage{
genHTML(x, outfile)
}
\arguments{
\item{x}{data}

\item{outfile}{A string denoting the location to put the output html
file.}
}
\value{
An html document
}
\description{
Generate Rmarkdown
}
\details{
Stuff to go here soon
}
\examples{
require(Rsynaptome)
set.seed(13)
n <- 100
x <- data.frame(v1 = v1 <<- sort(runif(n)), 
                v2 = v2 <<- sort(rnorm(n)), 
                v3 = v3 <<- ifelse(v1 > 0, v1, rnorm(1)),
                v4 = v4 <<- ifelse(v2 < 0, v2, rnorm(1)))
genHTML(x, "example.html")
}
\seealso{
\code{\link[rmarkdown]{render}}
}

