% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{genHTML}
\alias{genHTML}
\title{Generate an Rmarkdown with exploratory plots}
\usage{
genHTML(x, outfile, use.plotly = TRUE, scale = FALSE, dmethod = "pca",
  nmethod = "samp")
}
\arguments{
\item{x}{data}

\item{outfile}{A string denoting the location to put the output html
file.}

\item{use.plotly}{A Boolean to specify if ggplotly is used.}

\item{scale}{A Boolean to specify whether column-wise scaling is
performed before analysis.}

\item{dmethod}{a string denoting the method to use for dimension
reduction}

\item{nmethod}{a string denoting the method to use for sample
reduction}
}
\value{
An html document via RMarkdown.
}
\description{
Generate an Rmarkdown with exploratory plots
}
\details{
Generates an html file of various exploratory plots via
RMarkdown. "There is no excuse for failing to plot and look." ~ J. W.
Tukey (Exploratory Data Analysis, p 43)
}
\examples{
require(meda)
x <- iris[, -5]
y <- data.frame(matrix(rnorm(1e3*24, mean = rep(c(1,4,2,5), each =
1e3), sd = 0.25), ncol = 24))
load("~/neurodata/synaptome-stats/Code/cleanDataWithAttributes.RData")

outfile.1 <- paste0('~/Desktop/ex1.html')
outfile.2 <- paste0('~/Desktop/ex2.html')
outfile.3 <- paste0('~/Desktop/ex3.html')
use.plotly <- FALSE
scale <- FALSE
print("Now run 'genHTML(x, outfile, use.plotly, scale)'")
\dontrun{
x <- ex1[, 1:24, with = FALSE]
genHTML(x, outfile.1, use.plotly, scale)
genHTML(y, outfile.2, use.plotly, scale)
genHTML(data01[, 1:24, with = FALSE], outfile.3, use.plotly, scale)
p.heat(data01, FALSE, dmethod = "samp", nmethod = "kmpp")
p.heat(iris[, -5], FALSE, dmethod = "cur", nmethod = "kmpp")
cp <- comp(iris[, -5], dir = 2, nmethod='kmpp', nnum = 75)
}

}
\seealso{
\code{\link[rmarkdown]{render}}
}

