% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{genHTML}
\alias{genHTML}
\title{Generate an Rmarkdown with exploratory plots}
\usage{
genHTML(x, outfile, use.plotly = FALSE, scale = FALSE, dmethod = "cur",
  nmethod = "cur", truth = NULL, colCol = NULL, colRow = NULL)
}
\arguments{
\item{x}{data}

\item{outfile}{A string denoting the location to put the output html
file.}

\item{use.plotly}{A Boolean to specify if ggplotly is used.}

\item{scale}{A Boolean to specify whether column-wise scaling is
performed before analysis.}

\item{dmethod}{a string denoting the method to use for dimension
reduction}

\item{nmethod}{a string denoting the method to use for sample
reduction}

\item{truth}{an n length vector of integer values corresponding to
the true classes of the data.}

\item{colCol}{an n length vector of column colors.}

\item{colRow}{an n length vector of column colors.}
}
\value{
An html document via RMarkdown.
}
\description{
Generate an Rmarkdown with exploratory plots
}
\details{
Generates an html file of various exploratory plots via
RMarkdown. "There is no excuse for failing to plot and look." ~ J. W.
Tukey (Exploratory Data Analysis, p 43)
}
\examples{
require(meda)
dat <- iris[, -5]
truth <- as.numeric(iris[, 5])
featCol <- c("red", "purple", "darkgreen")
#p.heat(dat)
p.violin(dat)
p.outlier(dat)
do.call(corrplot, p.cor(dat))
p.cumvar(dat)
p.pairs(dat)
out <- p.bic(dat)
p.mclust(out$dat, out$bicO)
p.hmclust(dat, truth = iris[,5]) 

}
\seealso{
\code{\link[rmarkdown]{render}}
}

