% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{genHTML}
\alias{genHTML}
\title{Generate an Rmarkdown with exploratory plots}
\usage{
genHTML(x, outfile, use.plotly = FALSE, scale = FALSE, dmethod = "pca",
  nmethod = "samp")
}
\arguments{
\item{x}{data}

\item{outfile}{A string denoting the location to put the output html
file.}

\item{use.plotly}{A Boolean to specify if ggplotly is used.}

\item{scale}{A Boolean to specify whether column-wise scaling is
performed before analysis.}

\item{dmethod}{a string denoting the method to use for dimension
reduction}

\item{nmethod}{a string denoting the method to use for sample
reduction}
}
\value{
An html document via RMarkdown.
}
\description{
Generate an Rmarkdown with exploratory plots
}
\details{
Generates an html file of various exploratory plots via
RMarkdown. "There is no excuse for failing to plot and look." ~ J. W.
Tukey (Exploratory Data Analysis, p 43)
}
\examples{
require(meda)
dat <- iris[, -5]
p.heat(dat, FALSE, dmethod = "samp", nmethod = "kmpp")
p.violin(dat)
p.outlier(dat)
do.call(corrplot, p.cor(dat))
p.cumvar(dat)
p.pairs(dat)
p.bic(dat)
p.mclust(dat)

}
\seealso{
\code{\link[rmarkdown]{render}}
}

