% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcTree.R
\name{hmcTree}
\alias{hmcTree}
\title{Generate a binary hierarchical mclust tree}
\usage{
hmcTree(dat, maxDepth = 6)
}
\arguments{
\item{maxDepth}{maximum tree depth}

\item{data}{a data matrix}
}
\value{
binary hierarchical mclust classification output
}
\description{
Generate a binary hierarchical mclust tree
}
\details{
BIC is run for k = {1,2}, if k = 2 then each node is
propagated down the tree.  If k = 1, then that node is frozen.
}
\examples{
dat <- iris[, -5]
truth <- iris[, 5]
hmcL <- hmcTree(dat)
plot(as.dendrogram(L))
}

