% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmpp.R, R/zgElbows.R
\name{kmpp}
\alias{kmpp}
\title{K-Means++ initialization to K-means}
\usage{
kmpp(x, k = 2)

kmpp(x, k = 2)
}
\arguments{
\item{x}{The data formatted as a matrix.}

\item{k}{The number of clusters desired. \code{k < nrows(x) || length(x)}}

\item{x}{An input vector such as sdev returned from prcomp.}

\item{n}{The number of elbows to be returned.}

\item{th}{A threshold.}

\item{plot}{Boolean for output plot.}
}
\value{
An object of class `'kmeans''.
}
\description{
K-Means++ initialization to K-means

Dimension selection via Zhu adn Ghodsi
}
\details{
A set of centers are calculated as suggested by
Arthur and Vassilvitskii and used to initialize the kmeans algorithm.
The first center point is chosen at random.  Each subsequent center
is chosen with probability \eqn{P = D(x')^2/( \sum_{x \in X} D(x)^2)} (see references)


}
\examples{
set.seed(13)
z <- c(0,4,12,16)[sample(4, 1e3, replace=TRUE)]
x <- rnorm(1e3, mean = z, sd = 1)
kdf <- kmpp(x, 4)
par(mfrow = c(2,1))
plot(x, as.factor(z), col = as.factor(z), pch = '|')
title('Original Data')
plot(x, kdf$cluster, col = kdf$cluster, pch = '|')
title('Clustered Data')


set.seed(23)
s1 <- sample(4, 1e3, replace = TRUE)
z <- matrix(c(0,4,0,4,0,0,4,4), ncol=2)[s1,]
xy <- data.frame(x = rnorm(1e3, mean = z[,1]), y = rnorm(1e3, mean=z[,2]))
kdf2 <- kmpp(xy,4)
par(mfrow = c(2,1))
opal <- palette(adjustcolor(palette(), alpha.f = 0.55))
plot(xy, col = s1, pch = 19)
title('Original Data')
plot(xy, col = kdf2$cluster, pch = 17)
title('Clustered Data')
palette('default')

set.seed(13)
z <- c(0,4,12,16)[sample(4, 1e3, replace=TRUE)]
x <- rnorm(1e3, mean = z, sd = 1)
kdf <- kmpp(x, 4)
par(mfrow = c(2,1))
plot(x, as.factor(z), col = as.factor(z), pch = '|')
title('Original Data')
plot(x, kdf$cluster, col = kdf$cluster, pch = '|')
title('Clustered Data')


set.seed(23)
s1 <- sample(4, 1e3, replace = TRUE)
z <- matrix(c(0,4,0,4,0,0,4,4), ncol=2)[s1,]
xy <- data.frame(x = rnorm(1e3, mean = z[,1]), y = rnorm(1e3, mean=z[,2]))
kdf2 <- kmpp(xy,4)
par(mfrow = c(2,1))
opal <- palette(adjustcolor(palette(), alpha.f = 0.55))
plot(xy, col = s1, pch = 19)
title('Original Data')
plot(xy, col = kdf2$cluster, pch = 17)
title('Clustered Data')
palette('default')

}
\references{
Arthur, David, and Sergei Vassilvitskii. 
'k-means++: The advantages of careful seeding.'
Proceedings of the eighteenth annual ACM-SIAM
symposium on Discrete algorithms. Society for Industrial and Applied
Mathematics, 2007.
\url{http://ilpubs.stanford.edu:8090/778/1/2006-13.pdf}

Arthur, David, and Sergei Vassilvitskii. 
'k-means++: The advantages of careful seeding.'
Proceedings of the eighteenth annual ACM-SIAM
symposium on Discrete algorithms. Society for Industrial and Applied
Mathematics, 2007.
\url{http://ilpubs.stanford.edu:8090/778/1/2006-13.pdf}
}
\seealso{
\code{\link[stats]{kmeans}, \link{bhkmpp}}

\code{\link[stats]{kmeans}, \link{bhkmpp}}
}

