% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/l.R
\name{l}
\alias{l}
\title{List observations}
\usage{
l(df, by = NULL, n = 20)
}
\arguments{
\item{df}{\code{df} should be a data frame.}

\item{by}{\code{by} should be a string indicating variables in the data frame.}

\item{n}{\code{n} should be a positive integer.}
}
\value{
\code{l(df)} returns a data frame.
}
\description{
\code{l} lists observations in a data frame, just like the \code{l} command in Stata.
}
\examples{
abdata = haven::read_dta("http://www.stata-press.com/data/r13/abdata.dta")
abdata \%>\% dplyr::select(id, year, wage) \%>\% l("id", n = 30)
}

