% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{mlocation}
\alias{mlocation}
\title{Generate location estimates}
\usage{
mlocation(dat, ccol = NULL, CI = 0.95)
}
\arguments{
\item{dat}{data}

\item{ccol}{colors for features.}

\item{CI}{specified confidence interval about mean.}
}
\value{
a list of ggplot objects, see details.
}
\description{
Generate location estimates
}
\examples{
dat <- iris[, -5]
loc <- mlocation(dat, ccol = 1:4)
plot(loc)

}
