% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.3dpca}
\alias{p.3dpca}
\title{Generate 3D pca of correlation matrix}
\usage{
p.3dpca(dat, colCol = NULL)
}
\arguments{
\item{dat}{data}

\item{colCol}{colors for the columns of the data matrix}
}
\value{
a 3d scatter plot of first three PCs.
}
\description{
Generate 3D pca of correlation matrix
}
\examples{
dat <- iris[, -5]
truth <- iris[, 5]
colCol <- c("red", "green", "blue", "purple")
p.3dpca(dat, colCol)
}

