% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.bic}
\alias{p.bic}
\title{Generate a BIC plot}
\usage{
p.bic(dat, print = FALSE, G = 1:10, modelNames = NULL)
}
\arguments{
\item{dat}{data}

\item{print}{boolean for printing and plotting output.}

\item{G}{Number of components to test in mclustBIC, default 1:10}

\item{modelNames}{a list of models for bic to use, defaults to all.}
}
\value{
A BIC plot and as a side-effect a list of data and bic
output.
}
\description{
Generate a BIC plot
}
\examples{
out <- p.bic(iris[, -5])
}

