% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.bic}
\alias{p.bic}
\title{Generate a BIC plot}
\usage{
p.bic(dat, timeLimit = 8 * 60, print = FALSE)
}
\arguments{
\item{dat}{data}

\item{timeLimit}{Time limit for bic computation.}

\item{print}{boolean for printing and plotting output.}
}
\value{
A BIC plot and as a side-effect
}
\description{
Generate a BIC plot
}
\details{
Uses getElbows from
\url{http://www.cis.jhu.edu/~parky/Synapse/getElbows.R}
}
\examples{
out <- p.bic(iris[, -5])
}

