% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.clusterCov}
\alias{p.clusterCov}
\title{Generate cluster covariance plots}
\usage{
p.clusterCov(mod)
}
\arguments{
\item{mod}{mclust output}
}
\value{
heatmap and line plot of cluster means
}
\description{
Generate cluster covariance plots
}
\examples{
dat <- iris[, -5]
out <- p.bic(dat)
truth <- iris[, 5]
tryCatch(md1 <- p.mclust(out$dat, out$bicO, truth = truth))
mod <- md1 
p.clusterCov(mod)

}

