% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.clusterMeans}
\alias{p.clusterMeans}
\title{Generate cluster parameter plots}
\usage{
p.clusterMeans(modMeans, ccol = "black", cf = NULL)
}
\arguments{
\item{modMeans}{means from a model (clusters in columns)}

\item{ccol}{colors for features}

\item{cf}{cluster fractions}
}
\value{
heatmap and line plot of cluster means
}
\description{
Generate cluster parameter plots
}
\examples{
dat <- iris[, -5]
truth <- iris[, 5]
L <- p.hmc(dat, truth = truth)
p <- p.clusterMeans(L$means)
grid.arrange(p[[1]], p[[2]], ncol = 2)
}

