% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.clusterMeans}
\alias{p.clusterMeans}
\title{Generate cluster parameter plots}
\usage{
p.clusterMeans(modMeans, ccol = "black")
}
\arguments{
\item{modMeans}{means from a model (clusters in columns)}

\item{ccol}{colors for features}
}
\value{
heatmap and line plot of cluster means
}
\description{
Generate cluster parameter plots
}
\examples{
dat <- iris[, -5]
out <- p.bic(dat)
truth <- iris[, 5]
tryCatch(md1 <- p.mclust(out$dat, out$bicO, truth = truth))
modMeans <- md1$parameters$mean 
p <- p.clusterMeans(modMeans)
grid.arrange(p[[1]], p[[2]], ncol = 2)
}

