% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.cor}
\alias{p.cor}
\title{Generate a correlation plot}
\usage{
p.cor(dat, colCol = NULL)
}
\arguments{
\item{dat}{data}

\item{colCol}{colors for column labels}
}
\value{
A correlation plot
}
\description{
Generate a correlation plot
}
\examples{
dat <- iris[, -5]
colCol <- c("green", "red", "blue", "purple")
do.call(corrplot, p.cor(dat, colCol))
}

