% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.heat}
\alias{p.heat}
\title{Generate a heatmap plot}
\usage{
p.heat(dat, use.plotly, dmethod = "samp", nmethod = "samp")
}
\arguments{
\item{dat}{data}

\item{use.plotly}{Boolean to use plotly}

\item{dmethod}{parameter passed to \code{\link{comp}}}

\item{nmethod}{parameter passed to \code{\link{comp}}}
}
\value{
A heatmap plot of the data compressed if necessary.
}
\description{
Generate a heatmap plot
}

