% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.heat}
\alias{p.heat}
\title{Generate a heatmap plot}
\usage{
p.heat(dat, ...)
}
\arguments{
\item{dat}{data}

\item{...}{used to send options to heatmaply}
}
\value{
A heatmap plot of the data compressed if necessary.
}
\description{
Generate a heatmap plot
}

