% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.hmclust}
\alias{p.hmclust}
\title{Generate binary hierarchical mclust output}
\usage{
p.hmclust(dat, truth = NULL)
}
\arguments{
\item{dat}{data that p.bic has been run on}

\item{truth}{true labels if any}
}
\value{
binary hierarchical mclust classification output
}
\description{
Generate binary hierarchical mclust output
}
\examples{
dat <- iris[, -5]
truth <- iris[,5]
L <- p.hmclust(dat, truth)
}

