% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.mclust}
\alias{p.mclust}
\title{Generate mclust output}
\usage{
p.mclust(dat, bic, truth = NULL, print = FALSE)
}
\arguments{
\item{dat}{data that p.bic has been run on}

\item{bic}{output from p.bic or \code{\link[mclust]{mclustBIC}}}

\item{truth}{true labels if any}

\item{print}{Boolean to print parameter estimates.}
}
\value{
mclust classification output
}
\description{
Generate mclust output
}
\examples{
dat <- iris[, -5]
out <- p.bic(dat)
truth <- iris[, 5]
tryCatch(md1 <- p.mclust(out$dat, out$bicO, truth = truth))
}

