% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.outlier}
\alias{p.outlier}
\title{Generate an outlier plot}
\usage{
p.outlier(dat, k = sqrt(dim(dat)[1]), ...)
}
\arguments{
\item{dat}{data}

\item{k}{number of neighbors if nrow(dat) in (1e4,1e5]}

\item{...}{Unused}
}
\value{
An outlier plot
}
\description{
Generate an outlier plot
}
\details{
For sample size <= 1e4 an outlier measure is calculated from the 
randomForest package, points with measure greater than 3 sd from the
mean are considred outliers.  Need to add case for n > 1e4.
}

