% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.outlier}
\alias{p.outlier}
\title{Generate an outlier plot}
\usage{
p.outlier(dat)
}
\arguments{
\item{dat}{data}
}
\value{
An outlier plot
}
\description{
Generate an outlier plot
}
\details{
For each datapoint the average distance of the sqrt(n)-nearest
neighbors is computed. Call this set of points `mdk`. 
Outliers are then points with value 3 standard
deviations away from the mean(`mdk`).
}

