% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.stackM}
\alias{p.stackM}
\title{Generate stacked level mean plot}
\usage{
p.stackM(tree, ccol = "black", centered = FALSE, maxDepth = Inf,
  depth = 5)
}
\arguments{
\item{tree}{the hmcTree object}

\item{ccol}{colors for feature labels}

\item{centered}{boolen that skips level one if data was centered}

\item{maxDepth}{maximum number of levels to plot}

\item{depth}{forces via copying levels to show down to depth.}
}
\value{
a stacked level mean plot
}
\description{
Generate stacked level mean plot
}
\examples{
dat <- scale(iris[, -5], center = TRUE, scale = FALSE)
L <- hmcTree(dat)
p.dend(L)
p <- p.stackM(L, centered = TRUE)
print(p)
}

