% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.stackM}
\alias{p.stackM}
\title{Generate stacked level mean plot}
\usage{
p.stackM(tree, ccol = "black", centered = FALSE)
}
\arguments{
\item{tree}{the hmcTree object}

\item{ccol}{colors for feature labels}

\item{centered}{boolen that skips level one if data was centered}
}
\value{
a stacked level mean plot
}
\description{
Generate stacked level mean plot
}
\examples{
dat <- iris[, -5]
L <- hmcTree(dat)
p <- p.stackM(L)
print(p)
}

