% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHTML.R
\name{p.violin}
\alias{p.violin}
\title{Generate violin/jitter plot of data}
\usage{
p.violin(dat, use.plotly)
}
\arguments{
\item{dat}{data}

\item{use.plotly}{Boolean to use plotly}
}
\value{
A heatmap plot
}
\description{
Generate violin/jitter plot of data
}

