% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1heat.R
\name{plot.d1heat}
\alias{plot.d1heat}
\title{Generate 1d heatmap plots}
\usage{
\method{plot}{d1heat}(x, ...)
}
\arguments{
\item{x}{an object of type d1heat}

\item{...}{unused}
}
\value{
a ggplot object
}
\description{
Generate 1d heatmap plots
}
